✨ 히스토리 페이지에서 마인드맵 배경이 회색으로 흐려지고, 인터렉션 불가
- useDragBackground 커스텀 훅 분리
- HistoryBackground
