+++ client/src/components/atoms/Background/index.tsx
@@ -2,16 +2,17 @@ import React from 'react';
 import { BackgroundDiv, TSize, TColor } from './style';
 
 interface IProps {
-  bgSize?: TSize;
+  bgSize: TSize;
   bgColor?: TColor;
+  zIndex?: number;
   children?: React.ReactNode;
   id?: string;
   className?: string;
 }
 
-const Background: React.FC<IProps> = ({ className, bgColor, bgSize, children, id }) => {
+const Background: React.FC<IProps> = ({ className, bgColor, bgSize, children, id, zIndex }) => {
   return (
-    <BackgroundDiv id={id} className={className} bgColor={bgColor} bgSize={bgSize}>
+    <BackgroundDiv id={id} className={className} bgColor={bgColor} bgSize={bgSize} zIndex={zIndex}>
       {children}
     </BackgroundDiv>
   );
