--- client/src/components/atoms/svgCurve/index.tsx
@@ -1,6 +1,0 @@
-const svgCurve = () => {
-  //from to curve
-  return <svg></svg>;
-};
-
-export default svgCurve;
